Refactor threading into ParallelValidator
This change is intended to simplify Validator to enable emitting results
in a format other than Violation.
