+++ pkg/gcv/validator.go
@@ -18,8 +18,6 @@ package gcv
 import (
 	"context"
 	"encoding/json"
-	"flag"
-	"runtime"
 
 	"github.com/forseti-security/config-validator/pkg/api/validator"
 	asset2 "github.com/forseti-security/config-validator/pkg/asset"
@@ -46,16 +44,8 @@ const (
 	ancestorsKey = "ancestors"
 )
 
-var flags struct {
-	workerCount int
-}
-
-func init() {
-	flag.IntVar(
-		&flags.workerCount,
-		"workerCount",
-		runtime.NumCPU(),
-		"Number of workers that Validator will spawn to handle validate calls, this defaults to core count on the host")
+type ConfigValidator interface {
+	ReviewAsset(ctx context.Context, asset *validator.Asset) ([]*validator.Violation, error)
 }
 
 // Validator checks GCP resource metadata for constraint violation.
@@ -78,7 +68,6 @@ type Validator struct {
 	// These rego dependencies should be packaged with the GCV deployment.
 	// Right now expected to be set to point to "//policies/validator/lib" folder
 	policyLibraryDir string
-	work             chan func()
 	gcpCFClient      *cfclient.Client
 	k8sCFClient      *cfclient.Client
 }
@@ -135,7 +124,7 @@ func newCFClient(
 }
 
 // NewValidatorFromConfig creates the validator from a config.
-func NewValidatorFromConfig(stopChannel <-chan struct{}, config *configs.Configuration) (*Validator, error) {
+func NewValidatorFromConfig(config *configs.Configuration) (*Validator, error) {
 	gcpCFClient, err := newCFClient(gcptarget.New(), config.GCPTemplates, config.GCPConstraints)
 	if err != nil {
 		return nil, errors.Wrap(err, "unable to set up GCP Constraint Framework client")
@@ -147,53 +136,21 @@ func NewValidatorFromConfig(stopChannel <-chan struct{}, config *configs.Configu
 	}
 
 	ret := &Validator{
-		work:        make(chan func(), flags.workerCount*2),
 		gcpCFClient: gcpCFClient,
 		k8sCFClient: k8sCFClient,
 	}
-
-	go func() {
-		<-stopChannel
-		glog.Infof("validator shutdown requested via stopChannel close")
-		close(ret.work)
-	}()
-
-	workerCount := flags.workerCount
-	glog.Infof("validator starting %d workers", workerCount)
-	for i := 0; i < workerCount; i++ {
-		go ret.reviewWorker(i)
-	}
-
 	return ret, nil
 }
 
 // NewValidator returns a new Validator.
 // By default it will initialize the underlying query evaluation engine by loading supporting library, constraints, and constraint templates.
 // We may want to make this initialization behavior configurable in the future.
-func NewValidator(stopChannel <-chan struct{}, policyPaths []string, policyLibraryPath string) (*Validator, error) {
+func NewValidator(policyPaths []string, policyLibraryPath string) (*Validator, error) {
 	config, err := NewValidatorConfig(policyPaths, policyLibraryPath)
 	if err != nil {
 		return nil, err
 	}
-	return NewValidatorFromConfig(stopChannel, config)
-}
-
-func (v *Validator) reviewWorker(idx int) {
-	glog.V(1).Infof("worker %d starting", idx)
-	for f := range v.work {
-		f()
-	}
-	glog.V(1).Infof("worker %d terminated", idx)
-}
-
-// AddData adds GCP resource metadata to be audited later.
-func (v *Validator) AddData(request *validator.AddDataRequest) error {
-	return errors.Errorf("Not supported")
-}
-
-type assetResult struct {
-	violations []*validator.Violation
-	err        error
+	return NewValidatorFromConfig(config)
 }
 
 // ReviewAsset reviews a single asset.
@@ -213,18 +170,6 @@ func (v *Validator) ReviewAsset(ctx context.Context, asset *validator.Asset) ([]
 
 	assetMapInterface := assetInterface.(map[string]interface{})
 	return v.ReviewUnmarshalledJSON(ctx, assetMapInterface)
-}
-
-func (v *Validator) handleReview(ctx context.Context, idx int, asset *validator.Asset, resultChan chan<- *assetResult) func() {
-	return func() {
-		resultChan <- func() *assetResult {
-			violations, err := v.ReviewAsset(ctx, asset)
-			if err != nil {
-				return &assetResult{err: errors.Wrapf(err, "index %d", idx)}
-			}
-			return &assetResult{violations: violations}
-		}()
-	}
 }
 
 // fixAncestry will try to use the ancestors array to create the ancestorPath
@@ -337,34 +282,4 @@ func (v *Validator) reviewGCPResource(ctx context.Context, asset map[string]inte
 		return nil, errors.Wrapf(err, "GCP target Constraint Framework review call failed")
 	}
 	return v.convertResponses(gcptarget.Name, asset, responses)
-}
-
-// Review evaluates each asset in the review request in parallel and returns any
-// violations found.
-func (v *Validator) Review(ctx context.Context, request *validator.ReviewRequest) (*validator.ReviewResponse, error) {
-	assetCount := len(request.Assets)
-	resultChan := make(chan *assetResult, flags.workerCount*2)
-	defer close(resultChan)
-
-	go func() {
-		for idx, asset := range request.Assets {
-			v.work <- v.handleReview(ctx, idx, asset, resultChan)
-		}
-	}()
-
-	response := &validator.ReviewResponse{}
-	var errs multierror.Errors
-	for i := 0; i < assetCount; i++ {
-		result := <-resultChan
-		if result.err != nil {
-			errs.Add(result.err)
-			continue
-		}
-		response.Violations = append(response.Violations, result.violations...)
-	}
-
-	if !errs.Empty() {
-		return response, errs.ToError()
-	}
-	return response, nil
 }
